requirements.txt
numpy
pandas
scipy
pyaml
colorama
